feat: refactor code of component input and separates in diferent files
