+++ project/modules/form/ts/input/class-list.ts
@@ -1,0 +1,12 @@
+export const listClassName = [
+	'className',
+	'hasError',
+	'errorMessage',
+	'children',
+	'icon',
+	'floating',
+	'label',
+	'password',
+	'loading',
+	'colorSpinner',
+];
