+++ project/modules/form/ts/input/context.ts
@@ -1,0 +1,4 @@
+import React from 'react';
+
+export const InputContext = React.createContext({} as any);
+export const useInputContext = () => React.useContext(InputContext);
